requirements.txt
setuptools~=58.0.4
pyyaml~=6.0
requests~=2.26.0
